+++ perun/utils/log.py
@@ -228,3 +228,22 @@ def failed(ending='\n'):
     print('[', end='')
     cprint("FAILED", 'red', attrs=['bold'])
     print(']', end=ending)
+
+
+def print_minor_version(minor_version):
+    """Helper function for printing minor version to the degradation output
+
+    Currently printed in form of:
+
+    * sha[:6]: desc
+
+    :param MinorVersion minor_version: informations about minor version
+    """
+    print("* ", end='')
+    cprint("{}".format(
+        minor_version.checksum[:6]
+    ), 'yellow', attrs=[])
+    print(": {}".format(
+        minor_version.desc.split("\n")[0].strip()
+    ))
+